chore(docs): remove info parameters
