+++ packages/react/src/components/ContentSwitcher/ContentSwitcher.mdx
@@ -1,4 +1,4 @@
-import { Story, Props, Source, Preview } from '@storybook/addon-docs/blocks';
+import { Story, Args, Source, Canvas } from '@storybook/addon-docs/blocks';
 import ContentSwitcher from '../ContentSwitcher';
 import Switch from '../Switch';
 
@@ -37,13 +37,13 @@ Content switchers allow users to toggle between two or more content sections
 within the same space on screen. Only one content section is shown at a time.
 Create Switch components for each section in the content switcher.
 
-<Preview>
+<Canvas>
   <Story id="contentswitcher--default" />
-</Preview>
+</Canvas>
 
 ## Component API
 
-<Props />
+<Args />
 
 ### ContentSwitcher `className`
