+++ packages/react/src/components/CopyButton/CopyButton-story.js
@@ -38,10 +38,3 @@ export default {
 };
 
 export const Default = () => <CopyButton {...props()} />;
-
-Default.parameters = {
-  info: {
-    text:
-      'The copy button can be used when the user needs to copy information, such as a code snippet, to their clipboard.',
-  },
-};
