+++ packages/react/src/components/FileUploader/FileUploader-story.js
@@ -163,14 +163,6 @@ export const _FileUploaderButton = () => (
 
 _FileUploaderButton.storyName = 'FileUploaderButton';
 
-_FileUploaderButton.parameters = {
-  info: {
-    text: `
-        The FileUploaderButton can be used as a standalone component if you do not need the extra UI that comes with FileUploader. The FileUploaderButton is used in FileUploader.
-      `,
-  },
-};
-
 export const _FileUploader = () => {
   return (
     <div className={`${prefix}--file__container`}>
@@ -181,40 +173,17 @@ export const _FileUploader = () => {
 
 _FileUploader.storyName = 'FileUploader';
 
-_FileUploader.parameters = {
-  info: {
-    text: `
-        The FileUploader components allow the user to upload any necessary files. This uses the FileUploaderButton and Filename components. Filename components will appear below the FileUploaderButton when files are added. Use the filenameStatus prop to control what icon appears in Filename ('edit', 'complete', or 'uploading').
-      `,
-  },
-};
-
 export const _FileUploaderItem = () => (
   <FileUploaderItem {...props.fileUploaderItem()} />
 );
 
 _FileUploaderItem.storyName = 'FileUploaderItem';
 
-_FileUploaderItem.parameters = {
-  info: {
-    text: `
-      <FileUploaderItem /> represents an item that has been uploaded to the file uploader component. Use the \`status\` prop to control which icon appears ('edit', 'complete', or 'uploading').
-    `,
-  },
-};
-
 export const _FileUploaderDropContainer = () => (
   <FileUploaderDropContainer {...props.fileUploaderDropContainer()} />
 );
 
 _FileUploaderDropContainer.storyName = 'FileUploaderDropContainer';
-
-_FileUploaderDropContainer.parameters = {
-  info: {
-    text:
-      '<FileUploaderDropContainer /> is a drag and drop file uploader which allows users to upload files via both the normal file selection dialog and by dragging and dropping files.',
-  },
-};
 
 export const DragAndDropUploadContainerExampleApplication = () =>
   require('./stories/drop-container').default(
@@ -224,12 +193,6 @@ export const DragAndDropUploadContainerExampleApplication = () =>
 DragAndDropUploadContainerExampleApplication.storyName =
   'Drag and drop upload container example application';
 
-DragAndDropUploadContainerExampleApplication.parameters = {
-  info: {
-    text: 'Example application with drag and drop file uploader',
-  },
-};
-
 export const Skeleton = () => (
   <div style={{ width: '500px' }}>
     <FileUploaderSkeleton />
@@ -237,11 +200,3 @@ export const Skeleton = () => (
 );
 
 Skeleton.storyName = 'skeleton';
-
-Skeleton.parameters = {
-  info: {
-    text: `
-Placeholder skeleton state to use when content is loading.
-`,
-  },
-};
