+++ packages/react/src/components/FluidForm/FluidForm-story.js
@@ -57,15 +57,3 @@ export const Default = () => (
     />
   </FluidForm>
 );
-
-Default.parameters = {
-  info: {
-    text: `
-        Forms are widely used to collect user input.
-
-        Form can have any number of react components enclosed within FormGroup component. FormGroup component
-        is a wrapper for legend and fieldset component.
-
-      `,
-  },
-};
