+++ packages/react/src/components/Form/Form-story.js
@@ -250,15 +250,3 @@ export const Default = () => (
     </Button>
   </Form>
 );
-
-Default.parameters = {
-  info: {
-    text: `
-        Forms are widely used to collect user input.
-
-        Form can have any number of react components enclosed within FormGroup component. FormGroup component
-        is a wrapper for legend and fieldset component.
-
-      `,
-  },
-};
