+++ packages/react/src/components/Icon/Icon-story.js
@@ -78,25 +78,9 @@ export const Default = () => (
   </div>
 );
 
-Default.parameters = {
-  info: {
-    text: `
-        Icons are used in the product to present common actions and commands. Modify the fill property to change the color of the icon. The name property defines which icon to display. For accessibility, provide a context-rich description with the description prop. For a full list of icon names, see https://www.carbondesignsystem.com/guidelines/iconography/library
-      `,
-  },
-};
-
 export const Skeleton = () => (
   <div>
     <IconSkeleton {...propsSkeleton} />
     <IconSkeleton {...propsSkeleton2} />
   </div>
 );
-
-Skeleton.parameters = {
-  info: {
-    text: `
-        Icons are used in the product to present common actions and commands. Modify the fill property to change the color of the icon. The name property defines which icon to display. For accessibility, provide a context-rich description with the description prop. For a full list of icon names, see https://www.carbondesignsystem.com/guidelines/iconography/library
-      `,
-  },
-};
