+++ packages/react/src/components/Link/Link-story.js
@@ -38,12 +38,3 @@ export default {
 };
 
 export const Default = () => <Link {...props()}>Link</Link>;
-
-Default.parameters = {
-  info: {
-    text: `
-            Links are typically used as a means of navigation either within the application, to a place outside, or to a resource.
-            For anything else, especially things that change data, you should be using a button.
-          `,
-  },
-};
