+++ packages/react/src/components/Loading/Loading-story.js
@@ -32,13 +32,3 @@ export default {
 export const Default = () => {
   return <Loading {...props()} className={'some-class'} />;
 };
-
-Default.parameters = {
-  info: {
-    text: `
-        Loading spinners are used when retrieving data or performing slow computations,
-        and help to notify users that loading is underway. The 'active' property is true by default;
-        set to false to end the animation.
-      `,
-  },
-};
