+++ packages/react/src/components/Modal/Modal-story.js
@@ -177,30 +177,12 @@ export const _Default = () => {
   );
 };
 
-_Default.parameters = {
-  info: {
-    text: `
-        Modals communicate information via a secondary window and allow the user to maintain the context of a particular task.
-        Use the Modal Wrapper component to encapsulate your Modal within a button.
-      `,
-  },
-};
-
 export const TitleOnly = () => {
   const { size, ...rest } = titleOnlyProps();
   return <Modal {...rest} size={size || undefined}></Modal>;
 };
 
 TitleOnly.storyName = 'Title only';
-
-TitleOnly.parameters = {
-  info: {
-    text: `
-      In "small" and "xs" modals size, the title is allowed to span multiple lines and be used for the main message.
-      It should be less than 3 lines of text. If more room is required then use the standard body copy format.
-    `,
-  },
-};
 
 export const TrapFocus = () => {
   const { size, ...rest } = props();
@@ -223,12 +205,4 @@ export const TrapFocus = () => {
       />
     </Modal>
   );
-};
-
-TrapFocus.parameters = {
-  info: {
-    text: `
-        Specify a selector for the primary element to focus when opening a modal.
-      `,
-  },
 };
