+++ packages/react/src/components/OverflowMenu/OverflowMenu-story.js
@@ -86,15 +86,6 @@ export const Basic = withReadme(OverflowREADME, () => (
 
 Basic.storyName = 'basic';
 
-Basic.parameters = {
-  info: {
-    text: `
-        Overflow Menu is used when additional options are available to the user and there is a space constraint.
-        Create Overflow Menu Item components for each option on the menu.
-      `,
-  },
-};
-
 export const WithLinks = withReadme(OverflowREADME, () => (
   <OverflowMenu {...props.menu()}>
     <OverflowMenuItem
@@ -140,17 +131,6 @@ export const WithLinks = withReadme(OverflowREADME, () => (
 
 WithLinks.storyName = 'with links';
 
-WithLinks.parameters = {
-  info: {
-    text: `
-        Overflow Menu is used when additional options are available to the user and there is a space constraint.
-        Create Overflow Menu Item components for each option on the menu.
-
-        When given \`href\` props, menu items render as <a> tags to facilitate usability.
-      `,
-  },
-};
-
 export const CustomTrigger = withReadme(OverflowREADME, () => (
   <OverflowMenu
     {...{
@@ -178,11 +158,3 @@ export const CustomTrigger = withReadme(OverflowREADME, () => (
 ));
 
 CustomTrigger.storyName = 'custom trigger';
-
-CustomTrigger.parameters = {
-  info: {
-    text: `
-        Sometimes you just want to render something other than an icon
-      `,
-  },
-};
