+++ packages/react/src/components/RadioButton/RadioButton-story.js
@@ -50,17 +50,6 @@ export default {
 
 export const Default = () => <RadioButton id="radio-1" {...radioProps()} />;
 
-Default.parameters = {
-  info: {
-    text: `
-            Radio buttons are used when a list of two or more options are mutually exclusive,
-            meaning the user must select only one option. The example below shows how the Radio Button component
-            can be used as an uncontrolled component that is initially checked by setting the defaultChecked property
-            to true. To use the component in a controlled way, set the checked property instead.
-          `,
-  },
-};
-
 export const Skeleton = () => (
   <div>
     <RadioButtonSkeleton />
@@ -68,11 +57,3 @@ export const Skeleton = () => (
 );
 
 Skeleton.storyName = 'skeleton';
-
-Skeleton.parameters = {
-  info: {
-    text: `
-        Placeholder skeleton state to use when content is loading.
-      `,
-  },
-};
