+++ packages/react/src/components/Search/Search-story.js
@@ -58,16 +58,6 @@ export default {
 
 export const Default = () => <Search {...props()} id="search-1" />;
 
-Default.parameters = {
-  info: {
-    text: `
-            Search enables users to specify a word or a phrase to find particular relevant pieces of content
-            without the use of navigation. Search can be used as the primary means of discovering content,
-            or as a filter to aid the user in finding content.
-          `,
-  },
-};
-
 export const DeprecatedCustomButtons = () => (
   <div style={{ display: 'flex' }}>
     <Search {...props()} id="search-1" />
@@ -78,14 +68,6 @@ export const DeprecatedCustomButtons = () => (
 
 DeprecatedCustomButtons.storyName = '[Deprecated] custom buttons';
 
-DeprecatedCustomButtons.parameters = {
-  info: {
-    text: `
-        You can control what set of buttons you want.
-      `,
-  },
-};
-
 export const Skeleton = () => (
   <div style={{ width: '200px' }}>
     <SearchSkeleton />
@@ -95,11 +77,3 @@ export const Skeleton = () => (
 );
 
 Skeleton.storyName = 'skeleton';
-
-Skeleton.parameters = {
-  info: {
-    text: `
-        Placeholder skeleton state to use when content is loading.
-      `,
-  },
-};
