+++ packages/react/src/components/Select/Select-story.js
@@ -86,17 +86,6 @@ export const Default = () => {
   );
 };
 
-Default.parameters = {
-  info: {
-    text: `
-        Select displays a list below its title when selected. They are used primarily in forms,
-        where a user chooses one option from a list. Once the user selects an item, the dropdown will
-        disappear and the field will reflect the user's choice. Create Select Item components for each
-        option in the list. The example below shows an enabled Select component with three items.
-      `,
-  },
-};
-
 export const Skeleton = () => (
   <div
     style={{ width: '300px' }}
@@ -110,11 +99,3 @@ export const Skeleton = () => (
 );
 
 Skeleton.storyName = 'skeleton';
-
-Skeleton.parameters = {
-  info: {
-    text: `
-        Placeholder skeleton state to use when content is loading.
-      `,
-  },
-};
