+++ packages/react/src/components/StructuredList/StructuredList-story.js
@@ -73,14 +73,6 @@ export const Simple = () => (
   </StructuredListWrapper>
 );
 
-Simple.parameters = {
-  info: {
-    text: `
-        Structured Lists group content that is similar or related, such as terms or definitions.
-      `,
-  },
-};
-
 export const Selection = () => {
   const structuredListBodyRowGenerator = (numRows) => {
     return Array.apply(null, Array(numRows)).map((n, i) => (
@@ -127,14 +119,6 @@ export const Selection = () => {
   );
 };
 
-Selection.parameters = {
-  info: {
-    text: `
-    Structured Lists with selection allow a row of list content to be selected.
-  `,
-  },
-};
-
 export const Skeleton = () => (
   <div style={{ width: '800px' }}>
     <StructuredListSkeleton />
@@ -143,11 +127,3 @@ export const Skeleton = () => (
 );
 
 Skeleton.storyName = 'skeleton';
-
-Skeleton.parameters = {
-  info: {
-    text: `
-        Placeholder skeleton state to use when content is loading.
-      `,
-  },
-};
