+++ packages/react/src/components/Tag/Tag-story.js
@@ -59,31 +59,11 @@ export const _Default = () => (
   </Tag>
 );
 
-_Default.parameters = {
-  info: {
-    text: `
-        Tags are used for items that need to be labeled, categorized, or organized using keywords that describe them.
-        The example below shows how the Tag component can be used. Each type has a default message describing the type,
-        but a custom message can also be applied.
-      `,
-  },
-};
-
 export const Filter = () => (
   <Tag className="some-class" {...props.filter()} filter>
     {text('Content (children)', 'This is a tag')}
   </Tag>
 );
-
-Filter.parameters = {
-  info: {
-    text: `
-        Tags are used for items that need to be labeled, categorized, or organized using keywords that describe them.
-        The example below shows how the Tag component can be used. Each type has a default message describing the type,
-        but a custom message can also be applied.
-      `,
-  },
-};
 
 export const Skeleton = () => (
   <div>
@@ -92,11 +72,3 @@ export const Skeleton = () => (
 );
 
 Skeleton.storyName = 'skeleton';
-
-Skeleton.parameters = {
-  info: {
-    text: `
-      Placeholder skeleton state to use when content is loading.
-      `,
-  },
-};
