+++ packages/react/src/components/Tile/Tile-story.js
@@ -104,25 +104,9 @@ export const Default = () => {
   return <Tile {...regularProps}>Default tile</Tile>;
 };
 
-Default.parameters = {
-  info: {
-    text: `
-        Default tile without any interactions
-      `,
-  },
-};
-
 export const Clickable = () => {
   const clickableProps = props.clickable();
   return <ClickableTile {...clickableProps}>Clickable Tile</ClickableTile>;
-};
-
-Clickable.parameters = {
-  info: {
-    text: `
-        Clickable tile
-      `,
-  },
 };
 
 export const MultiSelect = () => {
@@ -144,16 +128,6 @@ export const MultiSelect = () => {
 
 MultiSelect.storyName = 'Multi-select';
 
-MultiSelect.parameters = {
-  info: {
-    text: `
-        Selectable tile
-
-        Use this to select multiple tiles.
-      `,
-  },
-};
-
 export const Radio = () => {
   const radioProps = props.radio();
   return (
@@ -174,22 +148,6 @@ export const Radio = () => {
   );
 };
 
-Radio.parameters = {
-  info: {
-    text: `
-         The example below shows a Tile Group component with a default selected Tile.
-         Although you can set the checked prop on the Tile, when using the RadioTile component
-         as a child of the Tile Group, either set the defaultSelected or valueSelected which will
-         automatically set the selected prop on the corresponding RadioTile component.
-
-         Use defaultSelected when you want a tile to be selected initially, but don't need to set it
-         at a later time. If you do need to set it dynamically at a later time, then use the valueSelected property instead.
-
-         Use this to select one tile at a time.
-      `,
-  },
-};
-
 export const Expandable = () => (
   <ExpandableTile {...props.expandable()}>
     <TileAboveTheFoldContent>
@@ -200,11 +158,3 @@ export const Expandable = () => (
     </TileBelowTheFoldContent>
   </ExpandableTile>
 );
-
-Expandable.parameters = {
-  info: {
-    text: `
-        Expandable tile
-      `,
-  },
-};
