+++ packages/react/src/components/Toolbar/Toolbar-story.js
@@ -87,11 +87,3 @@ export const Default = () => (
     </ToolbarItem>
   </Toolbar>
 );
-
-Default.parameters = {
-  info: {
-    text: `
-      Toolbar stuff
-    `,
-  },
-};
