+++ packages/react/src/components/Tooltip/Tooltip-story.js
@@ -138,16 +138,6 @@ export const DefaultBottom = () => (
 
 DefaultBottom.storyName = 'default (bottom)';
 
-DefaultBottom.parameters = {
-  info: {
-    text: `
-        Interactive tooltip should be used if there are actions a user can take in the tooltip (e.g. a link or a button).
-        For more regular use case, e.g. giving the user more text information about something, use definition tooltip or icon tooltip.
-        By default, the tooltip will render above the element. The example below shows the default scenario.
-      `,
-  },
-};
-
 export const NoIcon = () => (
   <div style={containerStyles}>
     <Tooltip {...props.withoutIcon()}>
@@ -167,16 +157,6 @@ export const NoIcon = () => (
 );
 
 NoIcon.storyName = 'no icon';
-
-NoIcon.parameters = {
-  info: {
-    text: `
-        Interactive tooltip should be used if there are actions a user can take in the tooltip (e.g. a link or a button).
-        For more regular use case, e.g. giving the user more text information about something, use definition tooltip or icon tooltip.
-        By default, the tooltip will render with an information Icon. The example below shows the option to exclude the Icon.
-      `,
-  },
-};
 
 export const RenderCustomIcon = () => (
   <div style={containerStyles}>
@@ -198,16 +178,6 @@ export const RenderCustomIcon = () => (
 
 RenderCustomIcon.storyName = 'render custom icon';
 
-RenderCustomIcon.parameters = {
-  info: {
-    text: `
-        Interactive tooltip should be used if there are actions a user can take in the tooltip (e.g. a link or a button).
-        For more regular use case, e.g. giving the user more text information about something, use definition tooltip or icon tooltip.
-        By default, the tooltip will render with an information Icon. The example below shows the option to exclude the Icon.
-      `,
-  },
-};
-
 export const OnlyCustomIcon = () => (
   <div style={containerStyles}>
     <Tooltip {...props.customIconOnly()}>
@@ -227,16 +197,6 @@ export const OnlyCustomIcon = () => (
 );
 
 OnlyCustomIcon.storyName = 'only custom icon';
-
-OnlyCustomIcon.parameters = {
-  info: {
-    text: `
-        Interactive tooltip should be used if there are actions a user can take in the tooltip (e.g. a link or a button).
-        For more regular use case, e.g. giving the user more text information about something, use definition tooltip or icon tooltip.
-        By default, the tooltip will render with an information Icon. The example below shows the option to exclude the Icon.
-      `,
-  },
-};
 
 export const UncontrolledTooltip = () => <UncontrolledTooltipExample />;
