+++ packages/react/src/components/TooltipDefinition/TooltipDefinition-story.js
@@ -58,13 +58,3 @@ export const Default = () => (
 );
 
 Default.storyName = 'default';
-
-Default.parameters = {
-  info: {
-    text: `
-      Definition tooltip is for regular use case of tooltip, e.g. giving the user more text information about something, like defining a word.
-      This works better than the interactive tooltip in regular use cases because the info icon used in interactive tooltip can be repetitive when it’s shown several times on a page.
-      Definition tooltip does not use any JavaScript. If there are actions a user can take in the tooltip (e.g. a link or a button), use interactive tooltip.
-    `,
-  },
-};
