+++ packages/react/src/components/UnorderedList/UnorderedList-story.js
@@ -35,15 +35,6 @@ export const Default = () => (
 
 Default.storyName = 'default';
 
-Default.parameters = {
-  info: {
-    text:
-      'Lists consist of related content grouped together and organized ' +
-      'vertically. Unordered lists are used to present content of equal ' +
-      'status or value.',
-  },
-};
-
 export const Nested = () => (
   <UnorderedList>
     <ListItem>
@@ -65,12 +56,3 @@ export const Nested = () => (
 );
 
 Nested.storyName = 'nested';
-
-Nested.parameters = {
-  info: {
-    text:
-      'Lists consist of related content grouped together and organized ' +
-      'vertically. Unordered lists are used to present content of equal ' +
-      'status or value.',
-  },
-};
